v2: updates & refacto
