+++ package.json
@@ -31,6 +31,7 @@
     "@vitejs/plugin-react": "^4.3.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "vite": "^5.4.11"
+    "vite": "^5.4.11",
+    "zod": "^3.23.8"
   }
 }
