+++ src/validation/validation.ts
@@ -1,0 +1,5 @@
+import { Player, PlayerSchema } from "./schemas/playerSchema";
+
+export function validatePlayers(players: unknown): Player[] {
+  return PlayerSchema.array().parse(players);
+}
